+++ public/assets/utils/getUserFromLocalStorage.js
@@ -1,0 +1,7 @@
+const getUserFromLocalStorage = () => {
+  const userStringified = localStorage.getItem('logtrade:::user');
+  if (userStringified === null) return null;
+  return JSON.parse(userStringified);
+}
+
+export default getUserFromLocalStorage;