+++ public/assets/utils/localStorage.js
@@ -1,0 +1,8 @@
+export const saveItemLocalStorage = (key, value) => 
+  localStorage.setItem(`${key}`, `${value}`);
+
+export const getItemFromLocalStorage = (key) => 
+  localStorage.getItem(`${key}`);
+
+export const removeItemFromLocalStorage = (key) => 
+  localStorage.removeItem(`${key}`);