+++ public/assets/vendors/font-awesome/font-awesome.js
@@ -1,0 +1,2 @@
+// ref: https://stackoverflow.com/a/52456022
+import '@fortawesome/fontawesome-free/js/all.js';