+++ src/index.js
@@ -1,163 +1,53 @@
-import isEqual from 'lodash.isequal';
-import './vendors/index';
+import React from 'react';
+import ReactDOM from 'react-dom';
+import './index.css';
+import App from './App';
+import { createNewTradeLog, deleteTradeLog, fetchAllTradesByUserId, updateTradeLog } from './vendors/firebase/firebase.firestore';
+// import reportWebVitals from './reportWebVitals';
 
-import Router from "./utils/Router.js";
-import render from "./utils/render.js";
-
-import Home from "./components/Home";
-import Login from './components/Login';
-import Signup from './components/Signup';
-import getUniqueId from './utils/getUniqueId';
-import Dashboard from './components/Dashboard';
-import getUserFromLocalStorage from './utils/getUserFromLocalStorage';
-import { fetchAllTradesByUserId } from './vendors/firebase/firebase.firestore';
-
-const appId = getUniqueId();
-
-const onLoad = () => {};
-
-const styles = () => `
-  * {
-    box-sizing: border-box;
-    padding: 0;
-    margin: 0;
-  }
-  html {
-    font-size: 16px;
-  }
-  body {
-    width: 100vw;
-    height: 100vh;
-    font-family: 'Poppins', sans-serif;
-  }
-  main {
-    max-width: 1200px;
-    width: 100%;
-    height: 100%;
-    margin: 0 auto;
-    overflow: auto;
-  }
-`;
+// const userId = 'H8XTiCvnxOWojohm2DewzlBxrjG2';
+// const newTradeLog = {
+//   tradeId: 'sd7f7f8d57fd',
+//   closingPrice: '15.00',
+//   openingPrice: '10.00',
+//   stopLoss: '9.00',
+//   takeProfit: '15.00',
+//   date: '13',
+//   month: 'October',
+//   year: '2021',
+//   notes: '',
+//   numberOfShares: 2,
+//   ticker: 'GOOGL',
+// };
+// const updatedTradeLog = {
+//   tradeId: 'sd7f7f8d57fd',
+//   closingPrice: '15.00',
+//   openingPrice: '10.00',
+//   stopLoss: '9.00',
+//   takeProfit: '15.00',
+//   date: '13',
+//   month: 'October',
+//   year: '2021',
+//   notes: '',
+//   numberOfShares: 2,
+//   ticker: 'LOL',
+// };
+// const tradeLogId = 'sd7f7f8d57fd';
 
 
-const store = (initialState = {}) => {
-  let state = {...initialState};
-  return {
-    getState: () => {
-      return {...state};
-    },
-    setState: (cb) => {
-      const updatedState = {...state, ...cb(state)};
-      console.log({ updatedState, state });
-      if (!isEqual(updatedState, state)) {
-        state = {...updatedState};
-        render(
-          {
-            getState,
-            setState,
-          }, 
-          appId,
-          App, 
-          styles,
-          onLoad,
-          document.getElementById('root'),
-        );
-      }
-    }
-  };
-}
+// fetchAllTradesByUserId('H8XTiCvnxOWojohm2DewzlBxrjG2');
+// createNewTradeLog(newTradeLog, userId);
+// deleteTradeLog(tradeLogId, userId);
+// updateTradeLog(tradeLogId, updatedTradeLog, userId);
 
-const dateToday = new Date(`October 2, 2021`);
+ReactDOM.render(
+  // <React.StrictMode>
+  <App />,
+  // </React.StrictMode>,
+  document.getElementById('root')
+);
 
-const initialState = { 
-  isLoggedIn: getUserFromLocalStorage() !== null,
-  dateToday,
-  activeDate: dateToday,
-  tradeLogs: [],
-  user: getUserFromLocalStorage(),
-  isDailyTradeModalOpen: false,
-}
-
-// global state
-const { getState, setState } = store(initialState);
-
-const App = (props = {}) => {
-  const { getState } = props;
-  const { 
-    isLoggedIn, 
-    dateToday,
-    activeDate,
-    tradeLogs,
-    user,
-    newTrade,
-    isDailyTradeModalOpen,
-  } = getState();
-  const state = {
-    isLoggedIn,
-    dateToday,
-    activeDate,
-    tradeLogs,
-    user,
-    newTrade,
-    isDailyTradeModalOpen,
-  }
-
-  console.log('App rendering...', { ...getState() });
-
-  return Router(
-    [
-      { 
-        page: Login,
-        redirectRules: [
-          { 
-            rule: () => isLoggedIn,
-            redirectTo: '?page=dashboard',
-          },
-        ],
-        matchingQuery: 'login',
-        props: { ...props, ...state },
-      },
-      { 
-        page: Signup, 
-        redirectRules: [
-          { 
-            rule: () => isLoggedIn,
-            redirectTo: '?page=dashboard',
-          },
-        ],
-        matchingQuery: 'signup',
-        props: { ...props, ...state },
-      },
-      { 
-        page: Dashboard,
-        redirectRules: [
-          { 
-            rule: () => !isLoggedIn,
-            redirectTo: '?page=home',
-          },
-        ],
-        matchingQuery: 'dashboard',
-        props: { ...props, ...state },
-      },
-    ], 
-    'page', 
-    { 
-      page: Home, 
-      props: { ...props, ...state, app: 'LogTrade' },
-    }
-  );
-};
-
-// render(
-//   {
-//     getState, 
-//     setState,
-//   }, 
-//   appId,
-//   App, 
-//   styles,
-//   onLoad,
-//   document.getElementById('root'),
-// );
-
-fetchAllTradesByUserId('H8XTiCvnxOWojohm2DewzlBxrjG2');
+// If you want to start measuring performance in your app, pass a function
+// to log results (for example: reportWebVitals(console.log))
+// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
+// reportWebVitals();